requirements.txt
openai-whisper
ffmpeg
moviepy
torch
torchaudio
nlptutti
scipy
language_tool_python
transformers
accelerate